+++ vao_blender/__init__.py
@@ -1,3 +1,3 @@
 """VAO-Blender implementation package."""
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"